run nikos as child of max_rss

+++ tests/max-rss/scripts/download-clang.sh
@@ -7,5 +7,8 @@ MACHINE=$(uname -m)
 ARCH="${arch_mapping[${MACHINE}]}"
 
 CLANG_URL="https://dd-agent-omnibus.s3.amazonaws.com/llvm/clang-14.0.5.${ARCH}"
-wget $CLANG_URL -O ./tools/clang-14
-chmod 0755 ./tools/clang-14
+if [ ! -f ./tools/clang-14 ]
+then
+  wget $CLANG_URL -O ./tools/clang-14
+  chmod 0755 ./tools/clang-14
+fi
